online_inference/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
scikit-learn==0.24.1
dataclasses==0.6
pyyaml==5.3.1
marshmallow-dataclass==8.3.0
pandas==1.1.5
hydra-core==1.0.6
pytest
fastapi==0.63.0
uvicorn==0.13.4
